Server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.17.0
idna==3.10
platformdirs==4.3.6
psycopg2-binary==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.2
django-cors-headers
pillow
yt-dlp==2025.2.19